feat: selectJdbcTemplate 템플릿 메서드 적용
